+++ sha2/CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.11.1 (UNRELEASED)
+### Changed
+- Removed workaround for unaligned loads in `riscv-zknh` backend ([#879])
+- `riscv-zknh` no longer requires `zbkb` (or `zbb`) target feature ([#879])
+
+[#879]: https://github.com/RustCrypto/hashes/pull/879
+
 ## 0.11.0 (2026-03-25)
 ### Changed
 - Edition changed to 2024 and MSRV bumped to 1.85 ([#652])